// interfaces/pull.request.interface.ts
export interface PullRequest {
  repository:  Repository;
  actor:       Actor;
  pullrequest: Pullrequest;
}

export interface Actor {
  display_name: string;
  links:        ActorLinks;
  type:         string;
  uuid:         string;
  account_id:   string;
  nickname:     string;
}

export interface ActorLinks {
  self:   Avatar;
  html:   Avatar;
  avatar: Avatar;
}

export interface Avatar {
  href: string;
}

export interface Pullrequest {
  comment_count:       number;
  task_count:          number;
  type:                string;
  id:                  number;
  title:               string;
  description:         string;
  rendered:            Rendered;
  state:               string;
  merge_commit:        null;
  close_source_branch: boolean;
  closed_by:           null;
  author:              Actor;
  reason:              string;
  created_on:          Date;
  updated_on:          Date;
  destination:         Destination;
  source:              Destination;
  reviewers:           any[];
  participants:        any[];
  links:               { [key: string]: Avatar };
  summary:             Summary;
}

export interface Destination {
  branch:     Branch;
  commit:     Commit;
  repository: Project;
}

export interface Branch {
  name: string;
}

export interface Commit {
  type:  string;
  hash:  string;
  links: CommitLinks;
}

export interface CommitLinks {
  self: Avatar;
  html: Avatar;
}

export interface Project {
  type:       string;
  full_name?: string;
  links:      ActorLinks;
  name:       string;
  uuid:       string;
  key?:       string;
  slug?:      string;
}

export interface Rendered {
  title:       Summary;
  description: Summary;
}

export interface Summary {
  type:   string;
  raw:    string;
  markup: string;
  html:   string;
}

export interface Repository {
  type:       string;
  full_name:  string;
  links:      ActorLinks;
  name:       string;
  scm:        string;
  website:    null;
  owner:      Owner;
  workspace:  Project;
  is_private: boolean;
  project:    Project;
  uuid:       string;
  parent:     null;
}

export interface Owner {
  display_name: string;
  links:        ActorLinks;
  type:         string;
  uuid:         string;
  username:     string;
}